# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Istio-Dashboard API",
        "title": "Istio-Dashboard",
        "contact": {
            "name": "sHuXnHs",
            "email": "610087273@qq.com"
        },
        "license": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/kube/namespace/list": {
            "get": {
                "description": "获取所有命名空间",
                "tags": [
                    "kube"
                ],
                "summary": "获取所有命名空间",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
        "/project/list": {
            "get": {
                "description": "获取所有的网格",
                "tags": [
                    "project"
                ],
                "summary": "获取所有网格",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
        "/sidecar/cds/list": {
            "get": {
                "description": "获取边车的CDS(集群配置)",
                "tags": [
                    "sidecar"
                ],
                "summary": "获取边车的CDS(集群配置)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pod",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
        "/sidecar/check": {
            "get": {
                "description": "检查边车配置是否同步",
                "tags": [
                    "sidecar"
                ],
                "summary": "检查边车配置是否同步",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pod",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
        "/sidecar/eds/list": {
            "get": {
                "description": "获取边车的EDS(端点配置)",
                "tags": [
                    "sidecar"
                ],
                "summary": "获取边车的EDS(端点配置)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pod",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
        "/sidecar/lds/list": {
            "get": {
                "description": "获取边车的LDS(所有监听端口)",
                "tags": [
                    "sidecar"
                ],
                "summary": "获取边车的LDS(所有监听端口)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pod",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
        "/sidecar/rds/list": {
            "get": {
                "description": "获取边车的RDS(路由)",
                "tags": [
                    "sidecar"
                ],
                "summary": "获取边车的RDS(路由)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pod",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "err_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}